refactor(tsquery): extract PreparedQuerying and PreparedOverlay
