+++ crates/tsquery/src/precompute_pattern_predicate.rs
@@ -1,0 +1,107 @@
+use hyperast::types::HyperAST;
+
+#[derive(Default)]
+pub struct PreparedQuerying<Q, HAST, Acc>(Q, std::marker::PhantomData<(HAST, Acc)>);
+
+impl<'a, HAST, Acc> From<&'a crate::Query> for PreparedQuerying<&'a crate::Query, HAST, Acc> {
+    fn from(value: &'a crate::Query) -> Self {
+        Self(value, Default::default())
+    }
+}
+
+impl<Q, HAST, Acc> std::ops::Deref for PreparedQuerying<Q, HAST, &Acc> {
+    type Target = Q;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl<T, HAST, Acc> hyperast::tree_gen::Prepro<T> for PreparedQuerying<&crate::Query, HAST, Acc> {
+    const USING: bool = false;
+
+    fn preprocessing(&self, ty: T) -> Result<hyperast::scripting::Acc, String> {
+        unimplemented!()
+    }
+}
+
+impl<'s, TS, T, Acc> hyperast::tree_gen::More for PreparedQuerying<&crate::Query, (TS, T), Acc>
+where
+    TS: 'static
+        + Clone
+        + hyperast::types::ETypeStore<Ty2 = Acc::Type>
+        + hyperast::types::RoleStore<IdF = u16, Role = hyperast::types::Role>,
+    T: hyperast::types::WithRoles,
+    T: hyperast::types::Tree,
+    T::TreeId: Copy,
+    Acc: hyperast::types::Typed
+        + hyperast::tree_gen::WithRole<hyperast::types::Role>
+        + hyperast::tree_gen::WithChildren<T::TreeId>,
+    for<'c> &'c Acc: hyperast::tree_gen::WithLabel<L = &'c str>,
+{
+    type Acc = Acc;
+    type T = T;
+    type TS = TS;
+    const ENABLED: bool = true;
+    fn match_precomp_queries<
+        'a,
+        HAST: HyperAST<
+                'a,
+                IdN = <Self::T as hyperast::types::Stored>::TreeId,
+                TS = Self::TS,
+                T = Self::T,
+            > + std::clone::Clone,
+    >(
+        &self,
+        stores: HAST,
+        acc: &Acc,
+        label: Option<&str>,
+    ) -> hyperast::tree_gen::PrecompQueries {
+        if self.0.enabled_pattern_count() == 0 {
+            return Default::default();
+        }
+        let pos = hyperast::position::StructuralPosition::empty();
+
+        let cursor = crate::cursor_on_unbuild::TreeCursor::new(stores, acc, label, pos);
+        let qcursor = self.0.matches_immediate(cursor); // TODO filter on height (and visibility?)
+        let mut r = Default::default();
+        for m in qcursor {
+            assert!(m.pattern_index.to_usize() < 16);
+            r |= 1 << m.pattern_index.to_usize() as u16;
+        }
+        r
+    }
+}
+
+impl<'s, TS, T, Acc> hyperast::tree_gen::PreproTSG<'s>
+    for PreparedQuerying<&crate::Query, (TS, T), Acc>
+where
+    TS: 'static
+        + Clone
+        + hyperast::types::ETypeStore<Ty2 = Acc::Type>
+        + hyperast::types::RoleStore<IdF = u16, Role = hyperast::types::Role>,
+    T: hyperast::types::WithRoles,
+    T: hyperast::types::Tree,
+    T::TreeId: Copy,
+    Acc: hyperast::types::Typed
+        + hyperast::tree_gen::WithRole<hyperast::types::Role>
+        + hyperast::tree_gen::WithChildren<T::TreeId>,
+    for<'c> &'c Acc: hyperast::tree_gen::WithLabel<L = &'c str>,
+{
+    const GRAPHING: bool = false;
+    fn compute_tsg<
+        HAST: HyperAST<
+                's,
+                IdN = <Self::T as hyperast::types::Stored>::TreeId,
+                TS = Self::TS,
+                T = Self::T,
+            > + std::clone::Clone,
+    >(
+        &self,
+        _stores: HAST,
+        _acc: &Acc,
+        _label: Option<&str>,
+    ) -> Result<usize, String> {
+        Ok(0)
+    }
+}
